refactor(backend): split Rust modules by domain
