+++ src-tauri/src/lib.rs
@@ -19,14 +19,14 @@ pub fn run() {
             data::save_data,
             dialog::open_file_dialog,
             dialog::open_folder_dialog,
-            project::analyze_project,
-            project::delete_project,
-            project::detect_editor_command,
-            project::export_projects,
-            project::import_projects,
-            project::open_project,
-            project::open_remote_project,
-            project::read_project_license,
+            project::commands::analyze_project,
+            project::commands::delete_project,
+            project::editor::detect_editor_command,
+            project::commands::export_projects,
+            project::commands::import_projects,
+            project::editor::open_project,
+            project::editor::open_remote_project,
+            project::license::read_project_license,
             scanner::detect_cli_history_root_path,
             scanner::detect_jetbrains_config_root_path,
             scanner::detect_recent_editor_state_db_path,
