+++ src-tauri/src/recent/models.rs
@@ -1,0 +1,5 @@
+#[derive(Clone)]
+pub struct RecentProject {
+    pub path: String,
+    pub ide: Option<String>,
+}
