+++ src-tauri/src/webdav.rs
@@ -1,77 +1,16 @@
-use std::{
-    fs,
-    path::PathBuf,
-    sync::LazyLock,
-    time::{SystemTime, UNIX_EPOCH},
-};
+mod error;
+mod files;
+mod models;
+mod remote;
+mod sync;
 
-use regex::Regex;
-use reqwest::{Client, Method, StatusCode};
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
-use tauri::{AppHandle, Manager};
-use url::Url;
+use tauri::AppHandle;
 
-use crate::data::{data_file_path, write_data_file_safely, DataFileKind};
-
-static URL_CREDENTIALS_PATTERN: LazyLock<Regex> = LazyLock::new(|| {
-    Regex::new(r"://[^:/@\s]+:[^@/\s]+@").expect("hardcoded regex should be valid")
-});
-
-/// 从错误消息中遮蔽 `scheme://user:pass@host` 形式的凭据,
-/// 防止 reqwest 错误把 Basic Auth 信息透传给前端。
-fn sanitize_url_for_error(error_msg: &str) -> String {
-    URL_CREDENTIALS_PATTERN
-        .replace_all(error_msg, "://***:***@")
-        .to_string()
-}
-
-/// reqwest 错误统一脱敏转换。
-fn request_error(error: reqwest::Error) -> String {
-    sanitize_url_for_error(&error.to_string())
-}
-
-#[derive(Clone, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct WebDavConfig {
-    endpoint: String,
-    username: Option<String>,
-    password: Option<String>,
-    remote_path: Option<String>,
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct WebDavSyncResult {
-    success: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    message: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    error: Option<String>,
-}
-
-struct SyncFile {
-    kind: DataFileKind,
-    name: &'static str,
-    fallback: &'static str,
-}
-
-const SYNC_FILES: [SyncFile; 2] = [
-    SyncFile {
-        kind: DataFileKind::Settings,
-        name: "settings.json",
-        fallback: "{}",
-    },
-    SyncFile {
-        kind: DataFileKind::Projects,
-        name: "projects.json",
-        fallback: "[]",
-    },
-];
+use models::{sync_error, sync_success, WebDavConfig, WebDavSyncResult};
 
 #[tauri::command]
 pub async fn webdav_upload_data(app: AppHandle, config: WebDavConfig) -> WebDavSyncResult {
-    match upload_data(app, config).await {
+    match sync::upload_data(app, config).await {
         Ok(count) => sync_success(format!("Uploaded {count} files")),
         Err(error) => sync_error(error),
     }
@@ -79,7 +18,7 @@ pub async fn webdav_upload_data(app: AppHandle, config: WebDavConfig) -> WebDavS
 
 #[tauri::command]
 pub async fn webdav_pull_data(app: AppHandle, config: WebDavConfig) -> WebDavSyncResult {
-    match pull_data(app, config).await {
+    match sync::pull_data(app, config).await {
         Ok(count) => sync_success(format!("Pulled {count} files")),
         Err(error) => sync_error(error),
     }
@@ -87,384 +26,8 @@ pub async fn webdav_pull_data(app: AppHandle, config: WebDavConfig) -> WebDavSyn
 
 #[tauri::command]
 pub async fn webdav_test_connection(config: WebDavConfig) -> WebDavSyncResult {
-    match test_connection(config).await {
+    match remote::test_connection(config).await {
         Ok(()) => sync_success("WebDAV connection OK".to_string()),
         Err(error) => sync_error(error),
-    }
-}
-
-async fn test_connection(config: WebDavConfig) -> Result<(), String> {
-    validate_config(&config)?;
-
-    let client = Client::builder()
-        .timeout(std::time::Duration::from_secs(10))
-        .build()
-        .map_err(|error| error.to_string())?;
-    let response = with_auth(
-        client.request(Method::OPTIONS, remote_url(&config, None, None)?),
-        &config,
-    )
-    .send()
-    .await
-    .map_err(request_error)?;
-
-    if response.status().is_success() {
-        return Ok(());
-    }
-
-    Err(format!("WebDAV connection failed: {}", response.status()))
-}
-
-async fn upload_data(app: AppHandle, config: WebDavConfig) -> Result<usize, String> {
-    validate_config(&config)?;
-
-    let client = Client::builder()
-        .timeout(std::time::Duration::from_secs(30))
-        .build()
-        .map_err(|error| error.to_string())?;
-    create_remote_dir(&client, &config).await?;
-
-    let mut count = 0;
-    for file in SYNC_FILES {
-        let path = data_file_path(&app, file.kind)?;
-        let raw = read_or_default(path, file.fallback)?;
-        let content = prepare_upload_content(file.kind, &raw)?;
-        let url = remote_file_url(&config, file.name)?;
-        let request = with_auth(client.put(url).body(content), &config)
-            .header("Content-Type", "application/json");
-        let response = request.send().await.map_err(request_error)?;
-        if !response.status().is_success() {
-            return Err(format!(
-                "Failed to upload {}: {}",
-                file.name,
-                response.status()
-            ));
-        }
-        count += 1;
-    }
-
-    Ok(count)
-}
-
-async fn pull_data(app: AppHandle, config: WebDavConfig) -> Result<usize, String> {
-    validate_config(&config)?;
-
-    let client = Client::builder()
-        .timeout(std::time::Duration::from_secs(30))
-        .build()
-        .map_err(|error| error.to_string())?;
-    let mut pulled_files = Vec::new();
-    for file in SYNC_FILES {
-        let url = remote_file_url(&config, file.name)?;
-        let response = with_auth(client.get(url), &config)
-            .send()
-            .await
-            .map_err(request_error)?;
-        if response.status() == StatusCode::NOT_FOUND {
-            return Err(format!("Remote file not found: {}", file.name));
-        }
-        if !response.status().is_success() {
-            return Err(format!(
-                "Failed to pull {}: {}",
-                file.name,
-                response.status()
-            ));
-        }
-
-        let raw = response.text().await.map_err(request_error)?;
-        let content = prepare_pull_content(&app, file.kind, &raw)?;
-        pulled_files.push((file.kind, content));
-    }
-
-    backup_local_sync_files(&app)?;
-
-    let count = pulled_files.len();
-    for (kind, content) in pulled_files {
-        write_data_file_safely(&data_file_path(&app, kind)?, &content)?;
-    }
-
-    Ok(count)
-}
-
-fn backup_local_sync_files(app: &AppHandle) -> Result<PathBuf, String> {
-    let backup_root = app
-        .path()
-        .app_data_dir()
-        .map_err(|error| error.to_string())?
-        .join("backups")
-        .join(format!(
-            "webdav-pull-{}",
-            SystemTime::now()
-                .duration_since(UNIX_EPOCH)
-                .map_err(|error| error.to_string())?
-                .as_secs()
-        ));
-    fs::create_dir_all(&backup_root).map_err(|error| error.to_string())?;
-
-    for file in SYNC_FILES {
-        let source = data_file_path(app, file.kind)?;
-        if source.exists() {
-            fs::copy(source, backup_root.join(file.name)).map_err(|error| error.to_string())?;
-        }
-    }
-
-    Ok(backup_root)
-}
-
-async fn create_remote_dir(client: &Client, config: &WebDavConfig) -> Result<(), String> {
-    let remote_path = normalized_remote_path(config);
-    if remote_path.is_empty() {
-        return Ok(());
-    }
-
-    let mut current = String::new();
-    for part in remote_path.split('/').filter(|part| !part.is_empty()) {
-        if !current.is_empty() {
-            current.push('/');
-        }
-        current.push_str(part);
-        if remote_dir_exists(client, config, &current).await? {
-            continue;
-        }
-
-        let url = remote_collection_url(config, &current)?;
-        let response = with_auth(
-            client.request(
-                Method::from_bytes(b"MKCOL").map_err(|error| error.to_string())?,
-                url,
-            ),
-            config,
-        )
-        .send()
-        .await
-        .map_err(request_error)?;
-        let status = response.status();
-        if status.is_success() {
-            continue;
-        }
-        if matches!(
-            status,
-            StatusCode::METHOD_NOT_ALLOWED | StatusCode::CONFLICT
-        ) && remote_dir_exists(client, config, &current).await?
-        {
-            continue;
-        }
-        return Err(format!("Failed to create remote directory: {}", status));
-    }
-
-    Ok(())
-}
-
-async fn remote_dir_exists(
-    client: &Client,
-    config: &WebDavConfig,
-    remote_path: &str,
-) -> Result<bool, String> {
-    let response = with_auth(
-        client
-            .request(
-                Method::from_bytes(b"PROPFIND").map_err(|error| error.to_string())?,
-                remote_collection_url(config, remote_path)?,
-            )
-            .header("Depth", "0"),
-        config,
-    )
-    .send()
-    .await
-    .map_err(request_error)?;
-
-    let status = response.status();
-    if status.is_success() || status.as_u16() == 207 {
-        return Ok(true);
-    }
-    if matches!(status, StatusCode::NOT_FOUND | StatusCode::CONFLICT) {
-        return Ok(false);
-    }
-
-    Err(format!("Failed to check remote directory: {}", status))
-}
-
-fn prepare_upload_content(kind: DataFileKind, raw: &str) -> Result<String, String> {
-    let mut value = parse_json(raw)?;
-    if matches!(kind, DataFileKind::Settings) {
-        if let Some(object) = value.as_object_mut() {
-            object.remove("webdav");
-        }
-    }
-    to_pretty_json(&value)
-}
-
-fn prepare_pull_content(app: &AppHandle, kind: DataFileKind, raw: &str) -> Result<String, String> {
-    let mut value = parse_json(raw)?;
-    if matches!(kind, DataFileKind::Settings) {
-        let local_path = data_file_path(app, DataFileKind::Settings)?;
-        if local_path.exists() {
-            let local = read_or_default(local_path, "{}")?;
-            match parse_json(&local) {
-                Ok(local_json) => {
-                    if let Some(webdav) = local_json.get("webdav").cloned() {
-                        if let Some(object) = value.as_object_mut() {
-                            object.insert("webdav".to_string(), webdav);
-                        }
-                    }
-                }
-                Err(error) => {
-                    // 本地 settings 损坏时跳过 webdav 合并(凭据可能丢失,
-                    // 用户需重新配置),但记录原因便于排查。
-                    eprintln!(
-                        "[webdav] Local settings.json is corrupted, skipping webdav config merge: {error}"
-                    );
-                }
-            }
-        }
-    }
-    to_pretty_json(&value)
-}
-
-fn read_or_default(path: PathBuf, fallback: &str) -> Result<String, String> {
-    match fs::read_to_string(path) {
-        Ok(content) => Ok(content),
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => Ok(fallback.to_string()),
-        Err(error) => Err(error.to_string()),
-    }
-}
-
-fn validate_config(config: &WebDavConfig) -> Result<(), String> {
-    if config.endpoint.trim().is_empty() {
-        return Err("WebDAV URL is required".to_string());
-    }
-    remote_file_url(config, "settings.json").map(|_| ())
-}
-
-fn remote_file_url(config: &WebDavConfig, file_name: &str) -> Result<Url, String> {
-    remote_url(
-        config,
-        Some(&normalized_remote_path(config)),
-        Some(file_name),
-    )
-}
-
-fn remote_url(
-    config: &WebDavConfig,
-    remote_path: Option<&str>,
-    file_name: Option<&str>,
-) -> Result<Url, String> {
-    let endpoint = ensure_trailing_slash(config.endpoint.trim());
-    let mut url = Url::parse(&endpoint).map_err(|error| error.to_string())?;
-    {
-        let mut segments = url
-            .path_segments_mut()
-            .map_err(|_| "Invalid WebDAV URL".to_string())?;
-        if let Some(remote_path) = remote_path {
-            for segment in remote_path.split('/').filter(|segment| !segment.is_empty()) {
-                segments.push(segment);
-            }
-        }
-        if let Some(file_name) = file_name {
-            segments.push(file_name);
-        }
-    }
-    Ok(url)
-}
-
-fn remote_collection_url(config: &WebDavConfig, remote_path: &str) -> Result<Url, String> {
-    let endpoint = ensure_trailing_slash(config.endpoint.trim());
-    let mut url = Url::parse(&endpoint).map_err(|error| error.to_string())?;
-    {
-        let mut segments = url
-            .path_segments_mut()
-            .map_err(|_| "Invalid WebDAV URL".to_string())?;
-        for segment in remote_path.split('/').filter(|segment| !segment.is_empty()) {
-            segments.push(segment);
-        }
-        segments.push("");
-    }
-    Ok(url)
-}
-
-fn normalized_remote_path(config: &WebDavConfig) -> String {
-    config
-        .remote_path
-        .as_deref()
-        .unwrap_or("CodeNest")
-        .trim()
-        .trim_matches('/')
-        .to_string()
-}
-
-fn ensure_trailing_slash(value: &str) -> String {
-    if value.ends_with('/') {
-        value.to_string()
-    } else {
-        format!("{value}/")
-    }
-}
-
-fn with_auth(request: reqwest::RequestBuilder, config: &WebDavConfig) -> reqwest::RequestBuilder {
-    let username = config.username.as_deref().unwrap_or("").trim();
-    if username.is_empty() {
-        return request;
-    }
-    request.basic_auth(username, config.password.clone())
-}
-
-fn parse_json(raw: &str) -> Result<Value, String> {
-    serde_json::from_str::<Value>(raw).map_err(|error| error.to_string())
-}
-
-fn to_pretty_json(value: &Value) -> Result<String, String> {
-    serde_json::to_string_pretty(value).map_err(|error| error.to_string())
-}
-
-fn sync_success(message: String) -> WebDavSyncResult {
-    WebDavSyncResult {
-        success: true,
-        message: Some(message),
-        error: None,
-    }
-}
-
-fn sync_error(error: String) -> WebDavSyncResult {
-    WebDavSyncResult {
-        success: false,
-        message: None,
-        error: Some(error),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::sanitize_url_for_error;
-
-    #[test]
-    fn sanitize_url_for_error_masks_credentials() {
-        assert_eq!(
-            sanitize_url_for_error(
-                "error sending request for url (https://user:secret@dav.example.com/path)"
-            ),
-            "error sending request for url (https://***:***@dav.example.com/path)"
-        );
-    }
-
-    #[test]
-    fn sanitize_url_for_error_keeps_urls_without_credentials() {
-        let message = "error sending request for url (https://dav.example.com/path)";
-        assert_eq!(sanitize_url_for_error(message), message);
-    }
-
-    #[test]
-    fn sanitize_url_for_error_handles_multiple_urls() {
-        assert_eq!(
-            sanitize_url_for_error("from http://a:b@x.com to https://c:d@y.com"),
-            "from http://***:***@x.com to https://***:***@y.com"
-        );
-    }
-
-    #[test]
-    fn sanitize_url_for_error_handles_ipv6_hosts() {
-        assert_eq!(
-            sanitize_url_for_error("error at https://user:secret@[::1]:8080/path"),
-            "error at https://***:***@[::1]:8080/path"
-        );
     }
 }
